@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.0+6
 
 environment:
-  sdk: ">=2.17.2 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -51,21 +51,20 @@ dependencies:
   #firebase_app_installations: ^0.2.2
 
   provider: ^6.0.5
-  shared_preferences: ^2.1.0 # TODO remove
   sqflite: ^2.2.6
   path: ^1.8.2
   http: ^0.13.5
-  go_router: ^11.0.0
+  go_router: ^12.0.0
   d_chart: #^2.2.11
     path: ../d_chart
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   cached_network_image: ^3.2.3
-  flutter_local_notifications: ^15.1.1
+  flutter_local_notifications: ^16.1.0
   share_plus: ^7.1.0
-  flutter_widget_from_html: ^0.10.5+3
+  flutter_widget_from_html: ^0.14.3
   package_info_plus: ^4.1.0
-  octo_image: ^1.0.2
+  octo_image: ^2.0.0
   google_sign_in: ^6.0.2
   awesome_extensions: ^2.0.5
   tap_debouncer: ^2.0.2
@@ -95,7 +94,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
   dart_code_metrics: ^5.7.6
 
 # For information on the generic Dart part of this file, see the
