@@ -3,11 +3,12 @@
  */
 
 import 'package:engelsburg_planer/main.dart';
-import 'package:engelsburg_planer/src/models/state/theme_state.dart';
-import 'package:engelsburg_planer/src/utils/firebase/analytics.dart';
+import 'package:engelsburg_planer/src/backend/database/state/app_state.dart';
+import 'package:engelsburg_planer/src/backend/database/state/theme_state.dart';
+import 'package:engelsburg_planer/src/services/firebase/analytics.dart';
+import 'package:engelsburg_planer/src/services/firebase/crashlytics.dart';
 import 'package:engelsburg_planer/src/utils/extensions.dart';
-import 'package:engelsburg_planer/src/utils/firebase/crashlytics.dart';
-import 'package:engelsburg_planer/src/utils/util.dart';
+import 'package:engelsburg_planer/src/utils/global_context.dart';
 import 'package:engelsburg_planer/src/view/routing/route_generator.dart';
 import 'package:engelsburg_planer/src/view/widgets/util/util_widgets.dart';
 import 'package:flutter/material.dart' hide Router;
@@ -15,18 +16,16 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
 
-import 'models/state/app_state.dart';
-
 /// The app itself
 class EngelsburgPlaner extends StatelessWidget {
-  const EngelsburgPlaner({Key? key}) : super(key: key);
+  const EngelsburgPlaner({super.key});
 
   @override
   Widget build(BuildContext context) {
     //Start app
     Analytics.logAppOpen();
     GlobalContext.firstContext = context;
-    InitializingPriority.needsContext.initialize();
+    InitializingPriority.context.initialize();
 
     return Consumer2<ThemeState, AppConfigState>(
       builder: (context, theme, config, _) {
