@@ -3,17 +3,16 @@
  */
 
 import 'package:awesome_extensions/awesome_extensions.dart';
-import 'package:engelsburg_planer/src/models/db/subjects.dart';
-import 'package:engelsburg_planer/src/models/state/user_state.dart';
-import 'package:engelsburg_planer/src/models/storage_adapter.dart';
+import 'package:engelsburg_planer/src/backend/database/nosql/model/subjects.dart';
+import 'package:engelsburg_planer/src/backend/database/nosql/base/document.dart';
+import 'package:engelsburg_planer/src/backend/database/state/user_state.dart';
 import 'package:engelsburg_planer/src/utils/extensions.dart';
-import 'package:engelsburg_planer/src/view/routing/page.dart';
 import 'package:engelsburg_planer/src/view/widgets/util/util_widgets.dart';
 import 'package:flutter/material.dart';
 
 /// Select a subject from a list. Returns document of subject or null of aborted.
 class SelectSubjectPage extends CompactStatelessWidget {
-  const SelectSubjectPage({Key? key}) : super(key: key);
+  const SelectSubjectPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -62,29 +61,29 @@ class SelectSubjectPage extends CompactStatelessWidget {
 
           return ListView.separated(
             padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 6),
-            itemBuilder: (context, index)  {
-                  var subject = subjects[index].data;
+            itemBuilder: (context, index) {
+              var subject = subjects[index].data;
 
-                  return ListTile(
-                    leading: Center(
-                      widthFactor: 1,
-                      child: SizedBox.square(
-                        dimension: 24,
-                        child: Container(
-                          decoration: BoxDecoration(
-                            color: subject?.parsedColor,
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                        ),
+              return ListTile(
+                leading: Center(
+                  widthFactor: 1,
+                  child: SizedBox.square(
+                    dimension: 24,
+                    child: Container(
+                      decoration: BoxDecoration(
+                        color: subject?.parsedColor,
+                        borderRadius: BorderRadius.circular(8),
                       ),
                     ),
-                    title: Text(
-                      subject?.parsedName(context) ?? "Subject deleted",
-                      textScaleFactor: 1.2,
-                    ),
-                    onTap: () => context.pop(result: subjects[index]),
-                  );
-                  },
+                  ),
+                ),
+                title: Text(
+                  subject?.parsedName(context) ?? "Subject deleted",
+                  textScaleFactor: 1.2,
+                ),
+                onTap: () => context.pop(result: subjects[index]),
+              );
+            },
             separatorBuilder: (context, index) => const Padding(
               padding: EdgeInsets.symmetric(horizontal: 8),
               child: Divider(height: 0, thickness: 1),
