@@ -2,22 +2,16 @@
  * Copyright (c) Paul Huerkamp 2023. All rights reserved.
  */
 
-import 'package:engelsburg_planer/src/models/api/substitutes.dart';
+import 'package:engelsburg_planer/src/backend/api/model/substitutes.dart';
 import 'package:engelsburg_planer/src/utils/extensions.dart';
-import 'package:engelsburg_planer/src/view/widgets/teacher_list_tile.dart';
-import 'package:engelsburg_planer/src/view/widgets/util/util_widgets.dart';
+import 'package:engelsburg_planer/src/view/pages/substitute/substitute_extended.dart';
 import 'package:flutter/material.dart';
 
-class SubstituteCard extends StatefulWidget {
-  const SubstituteCard({Key? key, required this.substitute}) : super(key: key);
+class SubstituteCard extends StatelessWidget {
+  const SubstituteCard({super.key, required this.substitute});
 
   final Substitute substitute;
 
-  @override
-  State<StatefulWidget> createState() => _SubstituteCardState();
-}
-
-class _SubstituteCardState extends State<SubstituteCard> {
   @override
   Widget build(BuildContext context) {
     final heroTag = StringUtils.randomAlphaNumeric(16);
@@ -29,335 +23,105 @@ class _SubstituteCardState extends State<SubstituteCard> {
       child: ClipRRect(
         borderRadius: BorderRadius.circular(8),
         child: Container(
-          color: _getTileColor(widget.substitute.type),
+          color: color(substitute.type),
           child: ListTile(
             minVerticalPadding: 8,
             onTap: () => context.pushPage(
-              ExtendedSubstituteCard(
-                substitute: widget.substitute,
+              ExtendedSubstitute(
+                substitute: substitute,
                 heroTag: heroTag,
               ),
             ),
             leading: Center(
               widthFactor: 1,
               child: Text(
-                widget.substitute.lesson!.toString(),
+                substitute.lesson!.toString(),
                 textScaleFactor: 1.8,
               ),
             ),
             title: Text(
-              widget.substitute.type.name(context),
+              substitute.type.name(context),
               textScaleFactor: 1.25,
             ),
-            subtitle: _buildText(),
+            subtitle: SummarizedSubstituteText(substitute: substitute),
           ),
         ),
       ),
     );
   }
 
-  Widget _buildText() {
+  static Color color(SubstituteType type) {
+    switch (type) {
+      case SubstituteType.canceled:
+        return Colors.red.shade700;
+      case SubstituteType.independentWork:
+        return Colors.purple.shade700;
+      case SubstituteType.roomSubstitute:
+        return Colors.lightBlueAccent.shade400;
+      case SubstituteType.care:
+        return Colors.green.shade600;
+      default:
+        return Colors.indigoAccent.shade700;
+    }
+  }
+}
+class SummarizedSubstituteText extends StatelessWidget {
+  const SummarizedSubstituteText({super.key, required this.substitute});
+
+  final Substitute substitute;
+  
+  @override
+  Widget build(BuildContext context) {
     return Wrap(
       children: [
         RichText(
           text: TextSpan(
-            text: widget.substitute.className,
+            text: substitute.className,
             style: DefaultTextStyle.of(context)
                 .style
                 .copyWith(color: DefaultTextStyle.of(context).style.color!.withOpacity(0.80)),
             children: [
-              TextSpan(text: widget.substitute.className == null ? '' : ' – '),
-              TextSpan(text: widget.substitute.subject),
+              TextSpan(text: substitute.className == null ? '' : ' – '),
+              TextSpan(text: substitute.subject),
               const TextSpan(text: ' ('),
               TextSpan(
-                  text: widget.substitute.substituteTeacher == null ||
-                          widget.substitute.substituteTeacher == '+'
+                  text: substitute.substituteTeacher == null ||
+                      substitute.substituteTeacher == '+'
                       ? ''
-                      : widget.substitute.substituteTeacher),
+                      : substitute.substituteTeacher),
               TextSpan(
-                  text: widget.substitute.substituteTeacher != null &&
-                          widget.substitute.substituteTeacher != '+' &&
-                          widget.substitute.teacher == null
+                  text: substitute.substituteTeacher != null &&
+                      substitute.substituteTeacher != '+' &&
+                      substitute.teacher == null
                       ? ')'
                       : ''),
               TextSpan(
-                  text: widget.substitute.substituteTeacher != null &&
-                          widget.substitute.substituteTeacher != '+' &&
-                          widget.substitute.teacher != null &&
-                          widget.substitute.substituteTeacher != widget.substitute.teacher
+                  text: substitute.substituteTeacher != null &&
+                      substitute.substituteTeacher != '+' &&
+                      substitute.teacher != null &&
+                      substitute.substituteTeacher != substitute.teacher
                       ? ' ${context.l10n.insteadOf} '
                       : ''),
               TextSpan(
-                  text: widget.substitute.teacher == widget.substitute.substituteTeacher
+                  text: substitute.teacher == substitute.substituteTeacher
                       ? ''
-                      : widget.substitute.teacher,
+                      : substitute.teacher,
                   style: const TextStyle(decoration: TextDecoration.lineThrough)),
-              TextSpan(text: widget.substitute.teacher != null ? ')' : ''),
+              TextSpan(text: substitute.teacher != null ? ')' : ''),
               TextSpan(
-                  text: widget.substitute.room == null ? '' : ' in ${widget.substitute.room!}'),
+                  text: substitute.room == null ? '' : ' in ${substitute.room!}'),
               TextSpan(
-                  text: widget.substitute.text == null || widget.substitute.text!.isEmpty
+                  text: substitute.text == null || substitute.text!.isEmpty
                       ? ''
-                      : ' – ${widget.substitute.text}'),
+                      : ' – ${substitute.text}'),
               TextSpan(
-                  text: widget.substitute.substituteOf == null
+                  text: substitute.substituteOf == null
                       ? ''
-                      : ' – ${widget.substitute.substituteOf}')
+                      : ' – ${substitute.substituteOf}')
             ],
           ),
         ),
       ],
-    );
-  }
-}
-
-Color _getTileColor(SubstituteType type) {
-  switch (type) {
-    case SubstituteType.canceled:
-      return Colors.red.shade700;
-    case SubstituteType.independentWork:
-      return Colors.purple.shade700;
-    case SubstituteType.roomSubstitute:
-      return Colors.lightBlueAccent.shade400;
-    case SubstituteType.care:
-      return Colors.green.shade600;
-    default:
-      return Colors.indigoAccent.shade700;
-  }
-}
-
-class ExtendedSubstituteCard extends CompactStatefulWidget {
-  const ExtendedSubstituteCard({
-    Key? key,
-    required this.substitute,
-    required this.heroTag,
-  }) : super(key: key);
-
-  final Substitute substitute;
-  final String heroTag;
-
-  @override
-  ExtendedSubstituteCardState createState() => ExtendedSubstituteCardState();
-}
-
-class ExtendedSubstituteCardState extends State<ExtendedSubstituteCard> {
-  @override
-  Widget build(BuildContext context) {
-    var subst = widget.substitute;
-    var start = Substitute.lessonStart(subst.lesson!);
-    var end = Substitute.lessonEnd(subst.lesson!);
-    String timeOfLessons = "$start - $end ${context.l10n.oclock}";
-
-    return Scaffold(
-      extendBodyBehindAppBar: true,
-      appBar: AppBar(
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.clear),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16),
-        child: ListView(
-          children: [
-            ListTile(
-              leading: Hero(
-                tag: widget.heroTag,
-                child: SizedBox.square(
-                  dimension: 20,
-                  child: Container(
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(4),
-                      color: _getTileColor(subst.type),
-                    ),
-                  ),
-                ),
-              ),
-              title: Align(
-                alignment: Alignment.centerLeft,
-                child: FittedBox(
-                  fit: BoxFit.scaleDown,
-                  child: Text(
-                    subst.type.name(context),
-                    style: const TextStyle(
-                      fontSize: 40,
-                      fontWeight: FontWeight.w300,
-                    ),
-                  ),
-                ),
-              ),
-              subtitle: Text(subst.date!.formatEEEEddMMToNow(context)),
-            ),
-            ListTile(
-              leading: const Icon(Icons.access_time),
-              dense: true,
-              title: Text(
-                subst.lesson!.toString(),
-                style: const TextStyle(fontSize: 18),
-              ),
-              subtitle: Text(timeOfLessons),
-            ),
-            ListTile(
-              leading: const Icon(Icons.class_),
-              dense: true,
-              title: Text(
-                subst.className!,
-                style: const TextStyle(fontSize: 18),
-              ),
-            ),
-            if (subst.substituteTeacher != null)
-              TeacherListTile(subst.substituteTeacher!),
-            if (subst.subject != null)
-              ListTile(
-                leading: const Icon(Icons.school),
-                dense: true,
-                title: Text(
-                  subst.subject!,
-                  style: const TextStyle(fontSize: 18),
-                ),
-              ),
-            if (subst.room != null)
-              ListTile(
-                leading: const Icon(Icons.room),
-                dense: true,
-                title: Text(
-                  subst.room!,
-                  style: const TextStyle(fontSize: 18),
-                ),
-              ),
-            if (subst.substituteOf != null)
-              ListTile(
-                leading: const Icon(Icons.event),
-                dense: true,
-                title: Text(
-                  "${context.l10n.substituteOf} ${subst.substituteOf!}",
-                  style: const TextStyle(fontSize: 18),
-                ),
-              ),
-            if (subst.text != null && subst.text!.isNotEmpty)
-              ListTile(
-                leading: const Icon(Icons.description),
-                dense: true,
-                title: Wrap(
-                  children: [
-                    Text(
-                      subst.text!,
-                      style: const TextStyle(fontSize: 18),
-                    ),
-                  ],
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class SubstituteMessageCard extends StatefulWidget {
-  const SubstituteMessageCard({Key? key, required this.substituteMessage}) : super(key: key);
-
-  final SubstituteMessage substituteMessage;
-
-  @override
-  SubstituteMessageCardState createState() => SubstituteMessageCardState();
-}
-
-class SubstituteMessageCardState extends State<SubstituteMessageCard> {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(10),
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(5),
-          color: Theme.of(context).textTheme.bodyLarge!.color!.withOpacity(0.15),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(10),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(10),
-                child: Text(
-                  widget.substituteMessage.date!.formatEEEEddMM(context),
-                  textScaleFactor: 2,
-                ),
-              ),
-              const Divider(height: 10, thickness: 5),
-              const SizedBox(height: 10),
-              Table(
-                defaultVerticalAlignment: TableCellVerticalAlignment.middle,
-                children: [
-                  if (widget.substituteMessage.absenceTeachers != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.absenceTeachers),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.absenceTeachers!),
-                        ),
-                      ],
-                    ),
-                  if (widget.substituteMessage.absenceClasses != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.absenceClasses),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.absenceClasses!),
-                        ),
-                      ],
-                    ),
-                  if (widget.substituteMessage.affectedClasses != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.affectedClasses),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.affectedClasses!),
-                        ),
-                      ],
-                    ),
-                  if (widget.substituteMessage.affectedRooms != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.affectedRooms),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.affectedRooms!),
-                        ),
-                      ],
-                    ),
-                  if (widget.substituteMessage.blockedRooms != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.blockedRooms),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.blockedRooms!),
-                        ),
-                      ],
-                    ),
-                  if (widget.substituteMessage.messages != null)
-                    TableRow(
-                      children: [
-                        Text(context.l10n.news),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 10),
-                          child: Text(widget.substituteMessage.messages!),
-                        ),
-                      ],
-                    ),
-                ],
-              )
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
