@@ -3,18 +3,18 @@
  */
 
 import 'package:awesome_extensions/awesome_extensions.dart';
-import 'package:engelsburg_planer/src/models/db/subjects.dart';
-import 'package:engelsburg_planer/src/models/db/tasks.dart';
-import 'package:engelsburg_planer/src/models/storage_adapter.dart';
-import 'package:engelsburg_planer/src/models/state/user_state.dart';
+import 'package:engelsburg_planer/src/backend/database/nosql/base/document.dart';
 import 'package:engelsburg_planer/src/utils/extensions.dart';
 import 'package:engelsburg_planer/src/view/pages/settings/subject/subject_select_page.dart';
 import 'package:engelsburg_planer/src/view/widgets/util/util_widgets.dart';
 import 'package:engelsburg_planer/src/view/widgets/util/wrap_if.dart';
 import 'package:flutter/material.dart';
+import 'package:engelsburg_planer/src/backend/database/nosql/model/subjects.dart';
+import 'package:engelsburg_planer/src/backend/database/nosql/model/tasks.dart';
+import 'package:engelsburg_planer/src/backend/database/state/user_state.dart';
 
 class ExtendedTask extends CompactStatefulWidget {
-  const ExtendedTask({Key? key, this.task, this.editing}) : super(key: key);
+  const ExtendedTask({super.key, this.task, this.editing});
 
   final Document<Task>? task;
   final bool? editing;
