@@ -2,16 +2,15 @@
  * Copyright (c) Paul Huerkamp 2023. All rights reserved.
  */
 
+import 'package:engelsburg_planer/src/backend/api/model/cafeteria.dart';
 import 'package:engelsburg_planer/src/backend/api/requests.dart';
-import 'package:engelsburg_planer/src/models/api/cafeteria.dart';
 import 'package:engelsburg_planer/src/utils/extensions.dart';
+import 'package:engelsburg_planer/src/view/widgets/util/api_future_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 
-import '../../widgets/util/api_future_builder.dart';
-
 class CafeteriaPage extends StatelessWidget {
-  const CafeteriaPage({Key? key}) : super(key: key);
+  const CafeteriaPage({super.key});
 
   @override
   Widget build(BuildContext context) {
